Refractor Command classes into separate modules
