+++ config/checkstyle/suppressions.xml
@@ -7,4 +7,5 @@
 <suppressions>
   <suppress checks="JavadocType" files=".*Test\.java"/>
   <suppress checks="MissingJavadocMethodCheck" files=".*Test\.java"/>
+  <suppress checks="CustomImportOder" files="../src/main/java/seedu/address/logic/parser/MasterDeckParser.java"/>
 </suppressions>
