+++ src/main/java/seedu/address/logic/commands/reviewcommands/NextCardCommand.java
@@ -1,28 +1,29 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.reviewcommands;
 
 import static java.util.Objects.requireNonNull;
 
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 
 /**
- * Marks the current card under review as wrong.
+ * Sets the next card as the card currently under review.
  */
-public class MarkWrongCommand extends Command {
+public class NextCardCommand extends Command {
 
-    public static final String COMMAND_WORD = ";";
+    public static final String COMMAND_WORD = "\\";
 
-    public static final String MESSAGE_SUCCESS = "You'll get it next time! Showing next flashcard."
+    public static final String MESSAGE_SUCCESS = "Skipped to next flashcard."
             + "\nEnter [ to flip card and show answer!"
             + "\nEnter ] to return to previous card."
             + "\nEnter \\ to skip to next card.";
-
-    public static final String MESSAGE_NO_MORE_NEXT_CARD = "Congrats! That was the last card.";
+    public static final String MESSAGE_NO_MORE_NEXT_CARD = "This is the last card.";
 
 
     @Override
     public CommandResult execute(Model model) {
         requireNonNull(model);
-        if (model.markWrong()) {
+        if (model.goToNextCard()) {
             return new CommandResult(MESSAGE_SUCCESS);
         } else {
             return new CommandResult(MESSAGE_NO_MORE_NEXT_CARD);
