+++ src/main/java/seedu/address/logic/parser/AddDeckCommandParser.java
@@ -1,6 +1,6 @@
 package seedu.address.logic.parser;
 
-import seedu.address.logic.commands.AddDeckCommand;
+import seedu.address.logic.commands.deckcommands.AddDeckCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.deck.Deck;
