+++ src/main/java/seedu/address/logic/parser/DeleteDeckCommandParser.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.DeleteDeckCommand;
+import seedu.address.logic.commands.deckcommands.DeleteDeckCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
