+++ src/main/java/seedu/address/logic/parser/MasterDeckParser.java
@@ -6,28 +6,28 @@ import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.AddDeckCommand;
 import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.DeleteDeckCommand;
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditDeckCommand;
-import seedu.address.logic.commands.EndReviewCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
-import seedu.address.logic.commands.FlipCardCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.MarkCorrectCommand;
-import seedu.address.logic.commands.MarkWrongCommand;
-import seedu.address.logic.commands.NextCardCommand;
-import seedu.address.logic.commands.PreviousCardCommand;
-import seedu.address.logic.commands.ReviewCommand;
-import seedu.address.logic.commands.SelectDeckCommand;
-import seedu.address.logic.commands.SetNumCardsPerReviewCommand;
-import seedu.address.logic.commands.UnselectDeckCommand;
+import seedu.address.logic.commands.cardcommands.AddCommand;
+import seedu.address.logic.commands.cardcommands.DeleteCommand;
+import seedu.address.logic.commands.cardcommands.EditCommand;
+import seedu.address.logic.commands.deckcommands.AddDeckCommand;
+import seedu.address.logic.commands.deckcommands.DeleteDeckCommand;
+import seedu.address.logic.commands.deckcommands.EditDeckCommand;
+import seedu.address.logic.commands.deckcommands.SelectDeckCommand;
+import seedu.address.logic.commands.deckcommands.UnselectDeckCommand;
+import seedu.address.logic.commands.reviewcommands.EndReviewCommand;
+import seedu.address.logic.commands.reviewcommands.FlipCardCommand;
+import seedu.address.logic.commands.reviewcommands.MarkCorrectCommand;
+import seedu.address.logic.commands.reviewcommands.MarkWrongCommand;
+import seedu.address.logic.commands.reviewcommands.NextCardCommand;
+import seedu.address.logic.commands.reviewcommands.PreviousCardCommand;
+import seedu.address.logic.commands.reviewcommands.ReviewCommand;
+import seedu.address.logic.commands.reviewcommands.SetNumCardsPerReviewCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
