+++ src/main/java/seedu/address/logic/parser/SetNumCardsPerReviewCommandParser.java
@@ -3,7 +3,7 @@ package seedu.address.logic.parser;
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
-import seedu.address.logic.commands.SetNumCardsPerReviewCommand;
+import seedu.address.logic.commands.reviewcommands.SetNumCardsPerReviewCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
