+++ src/test/java/seedu/address/logic/commands/AddDeckCommandTest.java
@@ -20,6 +20,7 @@ import javafx.collections.ObservableList;
 import javafx.util.Pair;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.deckcommands.AddDeckCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyMasterDeck;
