+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -11,6 +11,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.cardcommands.EditCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.MasterDeck;
 import seedu.address.model.Model;
